Backend/requirement.txt
Django==2.2.11
djangorestframework==3.11.0
pika==1.1.0
